refactor: getBestSkill as an actor method
